[UX][SREA Portal] Message group (#3427)
### Description
<!-- Provide a clear and concise description of the changes in this PR
-->

1. Group messages in the way that each group has the user messsage(s)
and its agent messages. In this way, we can easily figure out which
message should have header and footer and easy to calculate the copy
content. This also help grouping messages from reasoning API

- Before, we convert a message to a chatmessage with contents that
includes bunch of messages. But now we simply use the chat message group
which has userMessages array and agentMessages array in time ascending
order
- When loading old chat messages, we compare them with the oldest chat
message group to see if any message needs to be part of this group or
needs to start a new group. There are four scenarios:
- if the messsage is an agent message but the oldest message in the
message group is an user message, then we should not include this
message to the group, but creating a new group for this message.
- if the message is a user message but the oldest message in the message
group is an agent message, that means we found the user message for
those agent message, we should add this user message to the group
- if the message is an agent message and the oldest message in the
message group is also an agent message, that means they are all part of
the streaming session, thus add this message to the group
- if the message is a user message and the oldes message in the message
group is also a user message, that means the user aksed a message but
did not get any answer or cancelled the streaming before any agent
messages comes out. If this case, if two messages have different user id
or the timestamp is over 5 minutes, then the user message should be in a
new group. Otherwise add it in the userMessages array of the oldest
group

- Now we don't have to manually calculate which message needs header and
footer. Simple show header of the first message in the user message
array and first message in the agent message array. Show footer for the
last message in the agent message array. We don't have to recalculate
the copy content of all messages in the chatbox since we only care about
the agent messages within the group

- For getting user message history, we simply check all the user
messages of all groups instead of filtering every single message
manually

- I updated the scrolling implementation when landing on the thread
initially or sending a new messages, instead of relying on
useLayoutEffect or temporary message state, I simply just use setTimeout
or requestAnimationFrame when the messages are changed and the chatbox
is ready to scroll down. This way will avoid any early/late scroll due
to side effect like computing group messages. requestAnimationframe will
also make the scroll looks smoother


https://github.com/user-attachments/assets/4c197ae6-1b37-4e96-9310-a04bf7e5527e



https://github.com/user-attachments/assets/d3e5c6bf-167b-4300-b2d2-9bb6eea490a0

2. For streaming part, the previous implementation is that we memorize
if the thread has active streaming going on, if so, then we create a
default agent message as a placeholder for streaming message when
navigating away from the thread and coming back. The new implementation
uses the same logic but instead of creating a fake streaming message, we
will take the last message group as the streaming message group. The UI
looks much better as the previous implementation sometimes will create a
huge scroll space.

3. Remove the optional options for all fields of message object. In this
way, we can avoid the possibility of missing any relevant properties
when constructing a chat message. This is important when doing streaming
that we can have to figure out if it is pure text message or special
message as streaming chunk has all of them in 'text' field.




5. Unhide user scheduled task. Currently there is a bug that an
unexpected duplicated schduled task message is shown, Sanchit is working
on fixing it
<img width="523" height="107" alt="image"
src="https://github.com/user-attachments/assets/45597841-3476-43e2-9199-d59b0d6b1ad9"
/>


### Type of Change
<!-- Mark the appropriate option with an 'x', check all that apply -->
- [ ] Bug fix (non-breaking change that fixes an issue)
- [ ] New feature (non-breaking change that adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Documentation update
- [ ] Code refactoring
- [ ] Performance improvement
- [ ] Test coverage improvement
- [ ] Prompt change (requires approval from prompt author/owner)

### `Agent.Framework` Changes Checklist
<!-- If this PR includes changes to Agent.Framework, please complete
this section -->
- [ ] I have reviewed the [Agent.Framework contribution
guidelines](../src/Agent/Agent.Framework/CONTRIBUTING.md)
- [ ] The changes are appropriate for the framework (not business logic)
- [ ] The changes maintain the framework's domain-agnostic nature
- [ ] The changes provide clear extension points for consumers

### Risk
<!-- Mark the level of risk associated with this change -->
- [ ] High – Errors could cause MAJOR regression of many scenarios.
(Example: new large features or high-level infrastructure changes)
- [ ] Medium – Errors could cause regression of 1 or more scenarios.
(Example: somewhat complex bug fixes, small new features)
- [ ] Small – No issues are expected. (Example: Very small bugfixes,
string changes, or configuration settings changes)