3P SRE agent | Linux Fx Version | Fixing a bug where sitecontainers linuxfxversion being called as invalid (#3892)
This PR fixes the issue reported in 
- https://github.com/serverless-paas-balam/sreagent-runtime/issues/3624

**The bug:**
We had a list of supported runtime for which we need to run the
validation.
The bug was that when a site's linux fx version doesnt match the list,
we were calling the linux fx version as invalid.

**Fix:**
For unsupported runtime, assume that linux fx version is correct.

**Validation:**
Before fix:
<img width="715" height="265" alt="image"
src="https://github.com/user-attachments/assets/87e888d6-809b-43b1-af2d-3df7d77fad1e"
/>

Post fix:
Sitecontainers do not show up.